requirements.txt
playwright==1.49.0
playwright-stealth==1.0.6
python-telegram-bot==20.7
